+++ packages/app/electron-builder.yml
@@ -62,10 +62,8 @@ linux:
 extraResources:
   - from: '../../.env.production'
     to: '.env.production'
-  - from: 'build/icon_512x512.png'
-    to: 'icon.png'
-  - from: 'build/icon.ico'
-    to: 'icon.ico'
+  - from: 'src/static/icon-app.png'
+    to: 'icon-app.png'
 
 files:
   - '!**/*.map'
